move job/tasks-related code to job.py
