+++ python/init_db.py
@@ -22,6 +22,7 @@ from cm.logger import log
 from cm.models import UserProfile, DummyData, CheckLog, GroupCheckLog
 from cm.bundle import load_adcm
 from cm.config import SECRETS_FILE
+from cm.job import abort_all
 from cm.lock import unlock_all
 from cm.status_api import Event
 
@@ -71,6 +72,7 @@ def init():
         UserProfile.objects.create(login='admin')
     create_status_user()
     event = Event()
+    abort_all(event)
     unlock_all(event)
     clear_temp_tables()
     event.send_state()
